+++ integration_tests/test_bergermeer.py
@@ -22,7 +22,7 @@ def count_unique(arr):
 
 @pytest.mark.parametrize(
     "filename",
-    ["v2_bergermeer.sqlite"],
+    ["v2_bergermeer.gpkg"],
 )
 def test_integration(tmp_path, filename):
     shutil.copyfile(unittests_data_path / filename, tmp_path / filename)