+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog of threedigrid-builder
 
 - Try to get the extent of the 1D lines if extent of 1D nodes is None. Fixes issue with
   models that contain only embedded nodes.
+- Flip is_channel result to get correct is_closed value for connection nodes
 - Added reference to boundary lines and nodes to quarters administration.